+++ R/tmb_model_editors.R
@@ -25,6 +25,7 @@
 #' 
 #' @examples
 #' si = mp_tmb_library("starter_models", "si", package = "macpan2")
+#' print(si)
 #' 
 #' ## Update the mixing process to include 
 #' ## optional phenomenological heterogeneity.