+++ misc/dev/dev.cpp
@@ -107,8 +107,10 @@ enum macpan2_func
     , MP2_COS = 44 // fwrap,null: cos(x)
     , MP2_PRINT = 45 // fwrap,null: print(x)
     , MP2_TIME_VAR = 46 // fwrap,fail: time_var(x, change_points, change_pointer)
-    //, MP2_LOGISTIC = 47 // fwrap,null: logistic(x)
-    //, MP2_LOGIT = 48 // fwrap,null: logit(x)
+    , MP2_BINOM_SIM = 47 // fwrap,fail: rbinom(size, probability)
+    , MP2_EULER_MULTINOM_SIM = 48 // fwrap,fail: reulermultinom(size, rate, delta_t)
+    //, MP2_LOGISTIC = 48 // fwrap,null: logistic(x)
+    //, MP2_LOGIT = 49 // fwrap,null: logit(x)
 };
 
 enum macpan2_meth
@@ -890,18 +892,19 @@ public:
 
         // total number of time steps in the simulation loop
         int t_max = hist.size() - 2;
-
+        
+        // ---- Available Local Variables ----
         // Variables to use locally in 'macpan2 function' and
-        // 'macpan2 method' bodies -- scare quotes to clarify that
-        // these are not real functions and methods in either the
-        // c++ or r sense.
+        // 'macpan2 method' bodies -- these are not real functions and methods 
+        // in either the c++ or r sense.
         matrix<Type> m, m1, m2, m3, m4, m5;     // return values
         std::vector<int> v, v1, v2, v3, v4, v5; // integer vectors
-        vector<int> u;
+        vector<int> u; // FIXME: why not std::vector<int> here??
         matrix<Type> Y, X, A;
         std::vector<int> timeIndex; // for rbind_time and rbind_lag
         int doing_lag = 0;
-        Type sum, eps, var, by; // intermediate scalars
+        Type sum, eps, var, by, left_over, remaining_prop, p0; // intermediate scalars
+        Type delta_t; // for reulermultinom
         int rows, cols, lag, rowIndex, colIndex, matIndex, grpIndex, reps, cp, off, size;
         int sz, start, err_code, curr_meth_id;
         // size_t numMats;
@@ -1109,14 +1112,14 @@ public:
             // #' )
             // #' ```
             // #'
-            // #' To produce a simulation using these functions, one may
+            // #' To produce a simulation using these engine functions, one may
             // #' use \code{\link{simple_sims}}.
             // #'
             // #' ```
             // #' simple_sims(
             // #'   iteration_exprs = list(x ~ x - 0.9 * x),
             // #'   time_steps = 5,
-            // #'   x = 1
+            // #'   mats = list(x = 1)
             // #' )
             // #' ```
             // #'
@@ -2688,7 +2691,87 @@ public:
                     }
                 }
                 return m;
-
+            
+            case MP2_BINOM_SIM:
+                // rbinom(size, prob)
+                if (n != 2)
+                {
+                    SetError(MP2_BINOM_SIM, "rbinom needs two arguments: matrices with size and probability", row);
+                    return m;
+                }
+                rows = args[0].rows();
+                cols = args[0].cols();
+                v1.push_back(1);
+                args = args.recycle_to_shape(v1, rows, cols);
+                err_code = args.get_error_code();
+                // err_code = RecycleInPlace(args[1], rows, cols);
+                if (err_code != 0)
+                {
+                    SetError(err_code, "cannot recycle rows and/or columns because the input is inconsistent with the recycling request", row);
+                    return m;
+                }
+                m = matrix<Type>::Zero(rows, cols);
+                for (int i = 0; i < rows; i++)
+                {
+                    for (int j = 0; j < cols; j++)
+                    {
+                        m.coeffRef(i, j) = rbinom(args[0].coeff(i, j), args[1].coeff(i, j));
+                    }
+                }
+                return m;
+            
+            case MP2_EULER_MULTINOM_SIM:
+                //Rf_warning("experimental function reulermultinom");
+                // reulermultinom(size, rate, dt)
+                if (n == 3) {
+                    delta_t = args[2].coeff(0, 0);
+                } else {
+                    delta_t = 1.0;
+                }
+                if (args[0].rows() != 1 | args[0].cols() != 1) {
+                    //std::cout << "++++++" << std::endl;
+                    //std::cout << args[0] << std::endl;
+                    SetError(MP2_EULER_MULTINOM_SIM, "The first 'size' argument must be scalar.", row);
+                    return m;
+                }
+                if (args[1].cols() != 1) {
+                    //std::cout << "------" << std::endl;
+                    //std::cout << args[1] << std::endl;
+                    SetError(MP2_EULER_MULTINOM_SIM, "The second 'rate' argument must be a column vector", row);
+                }
+                sum = args[1].sum();
+                //std::cout << "sum of rates: " << sum << std::endl;
+                m = matrix<Type>::Zero(args[1].rows(), 1);  // multinomial probabilities
+                for (int i = 0; i < args[1].rows(); i++) {
+                    m.coeffRef(i, 0) = args[1].coeff(i, 0) * delta_t;
+                }
+                p0 = exp(-m.sum());
+                //std::cout << "prob(staying): " << p0 << std::endl;
+                p0 = (1 - p0) / sum;
+                //std::cout << "rate multiplier: " << p0 << std::endl;
+                for (int i = 0; i < args[1].rows(); i++) {
+                    m.coeffRef(i, 0) = p0 * args[1].coeff(i, 0);
+                }
+                //std::cout << "multinomial probabilities: " << m << std::endl;
+                
+                m1 = matrix<Type>::Zero(args[1].rows(), 1);  // multinomial outcomes
+                
+                left_over = args[0].coeff(0, 0);
+                remaining_prop = 1;
+                for (int i = 0; i < m1.rows(); i++) {
+                    //std::cout << "N = " << left_over << std::endl;
+                    //std::cout << "p = " << m.coeff(i, 0) << std::endl;
+                    //std::cout << "q = " << remaining_prop << std::endl;
+                    if (remaining_prop > 1e-8) {
+                        m1.coeffRef(i, 0) = rbinom(left_over, m.coeff(i, 0) / remaining_prop);
+                    }
+                    
+                    left_over = left_over - m1.coeff(i, 0);
+                    remaining_prop = remaining_prop - m.coeff(i, 0);
+                }
+                // m1.coeffRef(m1.rows() - 1, 0) = left_over;
+                return m1;
+                
             case MP2_ASSIGN:
                 // #' ## Assign
                 // #'
@@ -2894,6 +2977,32 @@ public:
                 }
                 return m;
 
+            // #' ## Print Matrix
+            // #' 
+            // #' Print out the value of a matrix.
+            // #' 
+            // #' ### Functions
+            // #'
+            // #' * `print(x)`
+            // #'
+            // #' ### Arguments
+            // #'
+            // #' * `x` -- Name of a matrix in the model.
+            // #'
+            // #' ### Return
+            // #'
+            // #' An \code{\link{empty_matrix}}.
+            // #'
+            // #' ### Examples
+            // #'
+            // #' ```
+            // #' simple_sims(
+            // #'      list(dummy ~ print(x), x ~ x / 2)
+            // #'    , time_steps = 10
+            // #'    , mats = list(x = 2)
+            // #' )
+            // #' ```
+            // #' 
             case MP2_PRINT:
                 std::cout << "printing matrix number " << index2mats[0] << " :" << std::endl;
                 std::cout << args[0] << std::endl;
@@ -2974,7 +3083,14 @@ public:
         int n = table_n[row];
         int x = table_x[row];
         int x1;
+        int x2;
+        int sz;
+        int nr;
+        int nc;
+        int size;
+        int start;
         matrix<Type> m;
+        matrix<Type> m1;
         std::vector<int> v1;
         std::vector<int> v2;
         int err_code;
@@ -2983,54 +3099,105 @@ public:
         // std::cout << "x: " << n << std::endl;
         switch (n)
         {
+        case 0:
+            valid_vars.m_matrices[x] = assignment_value;
+            return;
         case -1:
             Rf_error("trying to assign to a literal, which is not allowed");
         case -2:
             Rf_error("trying to assign to an engine method, which is not allowed");
         case -3:
             Rf_error("trying to assign to an integer vector, which is not allowed");
-        case 1:
-            Rf_error("assignment error -- TODO: be more specific");
-        case 0:
-            valid_vars.m_matrices[x] = assignment_value;
-            return;
-        case 2:
-            if (table_n[table_i[row] + 1] != -3)
-            {
-                Rf_error("indexing on the left-hand-side needs to be done using integer vectors");
-            }
-            v2.push_back(0);
-        case 3:
-            if (x + 1 != MP2_SQUARE_BRACKET)
-            {
-                Rf_error("square bracket is the only function allowed on the left-hand-side");
-            }
-            x1 = table_x[table_i[row]];
-
-            m = valid_vars.m_matrices[x1];
-            v1 = valid_int_vecs[table_x[table_i[row] + 1]];
-            if (n == 3)
-                v2 = valid_int_vecs[table_x[table_i[row] + 2]];
-
-            err_code = CheckIndices(m, v1, v2);
-            if (err_code)
-            {
-                Rf_error("assignment indexes are out of range");
-            }
-            assignment_value = RecycleToShape(assignment_value, v1.size(), v2.size());
-
-            for (int i = 0; i < v1.size(); i++)
-            {
-                for (int j = 0; j < v2.size(); j++)
-                {
-                    m.coeffRef(v1[i], v2[j]) = assignment_value.coeff(i, j);
+        }
+        // if we make it here we have a function on the left-hand-side.
+        // i.e. the number of arguments n > 0 (ignoring n < -3, which should 
+        // never get here because it would violate the spec). so now we need 
+        // to switch on the particular function being used. at most one 
+        // function can be used on the left-hand-side, and only particular ones 
+        // can be used as the switch statement shows
+        switch (x + 1) {
+            case MP2_SQUARE_BRACKET:
+                if (n == 3) { // two index vectors
+                    if (table_n[table_i[row] + 2] == -1) { // second index vector is a literal
+                        v2.push_back(CppAD::Integer(valid_literals[table_x[table_i[row] + 2]]));
+                        //Rf_error("indexing on the left-hand-side cannot be done using literals");
+                    } else if (table_n[table_i[row] + 2] != -3) { // second index vector is not an integer vector
+                        Rf_error("indexing on the left-hand-side needs to be done using integer vectors or literals");
+                    }
+                } else if (n == 2){ // one index vector
+                    v2.push_back(0);  // assume the second index vector is length-1 with a 0 (i.e. points to the first column)
+                } else { // too many index vectors
+                    Rf_error("only two index arguments allowed to square brackets on the left-hand-side");
                 }
-            }
-            valid_vars.m_matrices[x1] = m;
-            return;
-        default:
-            Rf_error("incorrect numbers of arguments");
-        } // switch (n)
+                if (table_n[table_i[row] + 1] == -1) { // first index vector is a literal
+                    v1.push_back(CppAD::Integer(valid_literals[table_x[table_i[row] + 1]]));
+                } else if (table_n[table_i[row] + 1] != -3) { // first index vector is not an integer vector or literal
+                    Rf_error("indexing on the left-hand-side needs to be done using integer vectors or literals");
+                } else { // first index is an integer vector
+                    v1 = valid_int_vecs[table_x[table_i[row] + 1]];
+                }
+                x1 = table_x[table_i[row]];
+
+                m = valid_vars.m_matrices[x1];
+                if (n == 3)
+                    v2 = valid_int_vecs[table_x[table_i[row] + 2]];
+    
+                err_code = CheckIndices(m, v1, v2);
+                if (err_code)
+                {
+                    Rf_error("assignment indexes are out of range");
+                }
+                assignment_value = RecycleToShape(assignment_value, v1.size(), v2.size());
+    
+                for (int i = 0; i < v1.size(); i++)
+                {
+                    for (int j = 0; j < v2.size(); j++)
+                    {
+                        m.coeffRef(v1[i], v2[j]) = assignment_value.coeff(i, j);
+                    }
+                }
+                valid_vars.m_matrices[x1] = m;
+                return;
+
+            case MP2_COMBINE:
+                // vectorize the right-hand-side so that it is a stacked 
+                // column vector
+                m = assignment_value;
+                size = m.rows() * m.cols();
+                m.resize(size, 1);
+                //std::cout << "assignment: " << m << std::endl;
+              
+                start = 0;
+                for (int i = 0; i < n; i++) {
+                    x2 = table_x[table_i[row]+i]; // index (in valid_vars) to ith argument of `c`
+                    //std::cout << "recipient: " << valid_vars.m_matrices[x2] << std::endl;
+                    sz = valid_vars.m_matrices[x2].rows() * valid_vars.m_matrices[x2].cols();
+                    if (sz == 0) sz = size / (n - i);  // heuristic
+                    //std::cout << "sz: " << sz << std::endl;
+                    //std::cout << "size: " << size << std::endl;
+                    //std::cout << "start: " << start << std::endl;
+                    if (size >= sz) {
+                        m1 = m.block(start, 0, sz, 1);
+                        nr = valid_vars.m_matrices[x2].rows();
+                        nc = valid_vars.m_matrices[x2].cols();
+                        if (nr == 0) nr = sz;
+                        if (nc == 0) nc = 1;
+                        m1.resize(nr, nc);
+                        //std::cout << "this replacement block: " << m1 << std::endl;
+                        valid_vars.m_matrices[x2] = m1;
+                        //std::cout << "recipient after: " << valid_vars.m_matrices[x2] << std::endl;
+                        size -= sz;
+                        start += sz;
+                    }
+                    else
+                        break;
+                }
+                // std::cout << "size at end: " << size << std::endl;
+                // std::cout << "size at end: " << size << std::endl;
+                return;
+        } // switch (x + 1)
+        
+        Rf_error("square bracket (e.g. x[i, j]) and concatenation (e.g. c(x, y, z)) are the only functions allowed on the left-hand-side");
     };
 };