+++ tests/testthat/test-by-running-library.R
@@ -11,6 +11,7 @@ test_that("all models in the library can run", {
       , "calibration_example.R"
       , package = "macpan2"
     )
-    if (file.exists(calibration_file)) source(calibration_file)
+    if (file.exists(calibration_file)) source(calibration_file, verbose= TRUE)
   }
 })
+