+++ NEWS.md
@@ -1,3 +1,32 @@
+# macpan2 1.2.0
+
+## Breaking Changes
+
+* No longer exporting `nlist`.
+
+## New Features
+
+* New stochasticity engine functions `rbinom` and `reulermultinom`.
+* `mp_tmb_fixef_cov` function for getting the covariance matrix of fixed effects.
+* `mp_trajectory_ensemble` function for summarizing random variation in trajectories.
+* Unpacking assignment in the engine. This means that you can do things like `c(x, y) ~ z`, where the values in `z` are assigned to elements in `x`, `y`, etc... in row-major order.
+* `to_name_pairs` function for returning all pairwise dot-concatenations of the elements of a character vector without dots.
+* `to_values` function for extracting the column from a data frame with only a single numerical column and turning numeric matrices and arrays with dimnames into a flattened numeric vector with labels produced by appropriately dot-concatenating the dimnames.
+
+## Bug Fixes
+
+* Several bugs related to input handling in `mp_tmb_calibrator` (#176)
+
+## Doc Fixes and Updates
+
+* New installation instructions for installing from `r-universe`.
+* New vignette: `real_data`.
+* Document `to_name`, `to_names`, `to_labels`, which handle naming of structured objects.
+* Document `print` function in the `?engine_functions`.
+* `simple_sims` example in `?engine_functions` now runs without error.
+* Help file examples for `mp_tmb_insert` and `mp_tmb_update`.
+* Readme/vignette examples better expose calls to `library` for the user.
+
 # macpan2 1.1.0
 
 ## Breaking Changes